libs/hive-config/pyproject.toml
[project]
name = "hive-config"
version = "0.0.1"
authors = [{ name = "Gary Benson", email = "gary@gbenson.net" }]
description = "Configuration management for Hive"
readme = "README.md"
dependencies = [
    "PyYAML",
]

[project.urls]
Homepage = "https://github.com/gbenson/hive/tree/main/libs/hive-config"
Source = "https://github.com/gbenson/hive"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "--cov=hive.config"

[tool.coverage.run]
omit = [
    "*/venv/*",
    "*/.venv/*",
]
